+++ src/query/executor/data_processing/transformations/append_vertices.rs
@@ -399,7 +399,7 @@ impl<S: StorageEngine + Send> ExecutorLifecycle for AppendVerticesExecutor<S> {
 }
 
 impl<S: StorageEngine + Send> ExecutorMetadata for AppendVerticesExecutor<S> {
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.base.id
     }
